Added Config control (default false) to reloading dependencies on startup (since cleaning up extension folder by default might be destructive for unaware clients that dont have copies or know what happened)

+++ OpenRPA/App.xaml.cs
@@ -236,18 +236,21 @@ namespace OpenRPA
                 }
 
                 // TODO: Clean extensions folder here
-                var extensionsPath = Path.Combine(Interfaces.Extensions.ProjectsDirectory, "extensions");
-                if (Directory.Exists(extensionsPath))
+                if (Config.local.restoreDependenciesOnStartup)
                 {
-                    foreach (var file in Directory.GetFiles(extensionsPath))
+                    var extensionsPath = Path.Combine(Interfaces.Extensions.ProjectsDirectory, "extensions");
+                    if (Directory.Exists(extensionsPath))
                     {
-                        try
+                        foreach (var file in Directory.GetFiles(extensionsPath))
                         {
-                            File.Delete(file);
-                        }
-                        catch (Exception ex)
-                        { 
-                            Log.Error("Could not clean extension: " + ex.ToString());
+                            try
+                            {
+                                File.Delete(file);
+                            }
+                            catch (Exception ex)
+                            {
+                                Log.Error("Could not clean extension: " + ex.ToString());
+                            }
                         }
                     }
                 }
